// apps/client-v2/src/api/generated/models/MovilDTO.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type MovilDTO = {
    /**
     * @description Identifier (Patente) of the movil
     * @type string
    */
    movil_id: string;
    /**
     * @description External system code
     * @type string | undefined
    */
    external_code?: string;
    /**
     * @description Internal inventory ID
     * @type string | undefined
    */
    inventory_id?: string;
    /**
     * @description Observations or notes
     * @type string | undefined
    */
    movil_observations?: string;
    /**
     * @description Type of movil (e.g., CIVIL, HIDRAULICA, RETIRO)
     * @type string
    */
    movil_type: string;
    /**
     * @description Operational state
     * @type string
    */
    movil_state: string;
    /**
     * @description Assigned conductor ID
     * @type integer
    */
    conductor_id?: number | null;
};